// src/types/Responses.d.ts
export interface RankingValueTier {
  rank: number;
  tier: string;
  value: number;
}
// ...
export interface UserDates {
  lastCitizenshipChangeAt: string;
  lastCompanyJoinedAt: string;
  lastConnectionAt: string;
  lastCountryMessageCheckAt: string;
  lastDailyRewardClaimedAt: string;
  lastEventsCheckAt: string;
  lastGlobalMessageCheckAt: string;
  lastHelpAskedAt: string;
  lastHiresAt: Array<string>;
  lastNotificationsCheckAt: string;
  lastSkillsResetAt: string;
  lastTakingControlAt: string;
  lastWorkAt: string;
  lastWorkOfferApplications: Array<string>;
}

export interface UserInfos {
  colorScheme: string;
  description: string;
  font: string;
  isPremium: boolean;
  premiumGiftsCount: number;
  premiumMonthsCount: number;
  vicePresidentOf: string;
}

export interface UserLeveling {
  availableSkillPoints: number;
  dailyXpLeft: number;
  freeReset: number;
  level: number;
  spentSkillPoints: number;
  totalSkillPoints: number;
  totalXp: number;
}

export interface UserRankings {
  userBounty: RankingValueTier;
  userCasesOpened: RankingValueTier;
  userDamages: RankingValueTier;
  userGemsPurchased: RankingValueTier;
  userLevel: RankingValueTier;
  userPremiumGifts: RankingValueTier;
  userPremiumMonths: RankingValueTier;
  userReferrals: RankingValueTier;
  userSubscribers: RankingValueTier;
  userTerrain: RankingValueTier;
  userWealth: RankingValueTier;
  weeklyUserDamages: RankingValueTier;
}

export interface UserSkillComputedStandard {
  equipment: number;
  level: number;
  limited?: unknown;
  total: number;
  totalAfterSoftCap?: unknown;
  value?: unknown;
  weapon?: unknown;
}

export interface UserSkillComputedAttack {
  ammoPercent: number;
  buffsPercent: number;
  debuffsPercent: number;
  equipment?: unknown;
  level: number;
  limited?: unknown;
  militaryRankPercent: number;
  total: number;
  totalAfterSoftCap?: unknown;
  value: number;
  weapon: number;
}

export interface UserSkillComputedCompany {
  equipment?: unknown;
  level: number;
  limited?: unknown;
  total: number;
  totalAfterSoftCap?: unknown;
  value: number;
  weapon?: unknown;
}

export interface UserSkillComputedCriticalChance {
  equipment?: unknown;
  level: number;
  limited?: unknown;
  total: number;
  totalAfterSoftCap?: unknown;
  value: number;
  weapon: number;
}

export interface UserSkillComputedPrecision {
  equipment: number;
  level: number;
  limited?: unknown;
  total: number;
  totalAfterSoftCap?: unknown;
  value: number;
  weapon?: unknown;
}

export interface UserSkillComputedBar {
  currentBarValue: number;
  equipment?: unknown;
  hourlyBarRegen: number;
  level: number;
  limited?: unknown;
  total: number;
  totalAfterSoftCap?: unknown;
  value: number;
  weapon?: unknown;
}

export interface UserSkills {
  armor: UserSkillComputedStandard;
  attack: UserSkillComputedAttack;
  companies: UserSkillComputedCompany;
  criticalChance: UserSkillComputedCriticalChance;
  criticalDamages: UserSkillComputedPrecision;
  dodge: UserSkillComputedStandard;
  energy: UserSkillComputedBar;
  entrepreneurship: UserSkillComputedBar;
  health: UserSkillComputedBar;
  hunger: UserSkillComputedBar;
  lootChance: UserSkillComputedCompany;
  management: UserSkillComputedCompany;
  precision: UserSkillComputedPrecision;
  production: UserSkillComputedBar;
}

export interface UserStats {
  damagesCount: number;
}
// ...
export type UserGetUserLiteResponse = {
  _id: string;
  avatarUrl: string;
  country: string;
  createdAt: string;
  dates: UserDates;
  infos: UserInfos;
  isActive: boolean;
  leveling: UserLeveling;
  militaryRank: number;
  mu: string;
  rankings: UserRankings;
  skills: UserSkills;
  stats: UserStats;
  username: string;
};